linux: use intptr_t where appropriate to fix 64-bit build
Review URL: http://codereview.chromium.org/553157

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37558 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/webplugin_impl.cc
@@ -355,7 +355,7 @@ void WebPluginImpl::didFinishLoadingFrameRequest(
     const WebURL& url, void* notify_data) {
   if (delegate_) {
     delegate_->DidFinishLoadWithReason(
-        url, NPRES_DONE, reinterpret_cast<int>(notify_data));
+        url, NPRES_DONE, reinterpret_cast<intptr_t>(notify_data));
   }
 }
 
@@ -367,7 +367,7 @@ void WebPluginImpl::didFailLoadingFrameRequest(
   NPReason reason =
       error.reason == net::ERR_ABORTED ? NPRES_USER_BREAK : NPRES_NETWORK_ERR;
   delegate_->DidFinishLoadWithReason(
-      url, reason, reinterpret_cast<int>(notify_data));
+      url, reason, reinterpret_cast<intptr_t>(notify_data));
 }
 
 // -----------------------------------------------------------------------------
